+++ zntrack/core/__init__.py
@@ -1,6 +1,1 @@
 """Core module of zntrack."""
-# from zntrack.core.base import Node
-# from zntrack.core.functions.decorator import NodeConfig, nodify
-# from zntrack.core.zntrackoption import ZnTrackOption
-#
-# __all__ = ["Node", "ZnTrackOption", "nodify", "NodeConfig"]
