--- zntrack/core/functions/__init__.py
@@ -1,1 +1,0 @@
-"""ZnTrack function decorator module."""
