+++ zntrack/core/node.py
@@ -1,3 +1,4 @@
+"""The Node class."""
 from __future__ import annotations
 
 import dataclasses
