+++ zntrack/fields/field.py
@@ -1,3 +1,4 @@
+"""The base class for all fields."""
 import abc
 import json
 import pathlib
