--- zntrack/interface/__init__.py
@@ -1,5 +1,0 @@
-"""Some basic interface functionality."""
-
-from zntrack.interface.base import DVCInterface
-
-__all__ = ["DVCInterface"]
