--- zntrack/metadata/__init__.py
@@ -1,5 +1,0 @@
-"""meta data collections such as execution times."""
-from .base import MetaData
-from .decorators import TimeIt
-
-__all__ = ["MetaData", "TimeIt"]
