--- zntrack/zn/__init__.py
@@ -1,7 +1,0 @@
-"""Description: zn.<option>.
-
-The following can be used to store e.g. metrics directly without
-defining and writing to a file explicitly. For more information on the individual methods
-see https://dvc.org/doc/command-reference/run#options
-
-"""
